requirement.txt
fastapi
uvicorn
pydantic
transformers
torch
faiss-cpu
langchain
langchain-classic
langchain-google-genai
openai
sentence-transformers
openai-whisper
numpy
scipy
python-multipart
pydantic-settings
accelerate
google-generativeai
datasets
huggingface-hub
kagglehub
supabase
gTTS
